Updated jquery.backgroundvideo.js
Modified to optionally accept an object as the path argument instead of a string. If you have to reference video files in different locations (for example on a Wordpress site where they were not uploaded at the same time and are in different subdirectories of wp-content/uploads) this can be useful. Object properties should be strings corresponding to the members of the types array.